fix: unreachable letter combinations in live-preview mode
Addresses issue #64. In live preview mode, when there are both markdown and live preview links, the letter combinations to jump to those links will sometimes be unreachable, for example both "a" and "aa" being present.

For an example of how this occurs, we'll assume there are three markdown links and three live preview links, and the characters used for link hints are "abcd":

1. First the three markdown links are found and three letters generated for them: ["a", "b", "c"
2. Next, the three live preview links are found and three letters generated for them: ["a", "b", "c"]
3. To disambiguate these links, we then generate a number of letters equal to the total number of links: ["b", "c", "d", "aa", "ab", "ac"]
4. These last three letters are used to overwrite the live preview links: ["aa", "ab", "ac"]
5. Finally, the markdown and live preview link arrays are combined, resulting in the bugged behaviour: ["a", "b", "c", "aa", "ab", "ac"]

I've fixed the bug by also overwriting the letters for the markdown links with the first three letters generated in step three. A more elegant fix would involve pulling the letter generation functionality out of the three link hint functions (`getPreviewLinkHints`, `getMDLinkHints` and `extractRegexBlocks`) so that the letters can be applied to the links later in a single step.