fix: prevent popovers displaying undefined in live-preview mode
Since `getLinkHintLetters` can only generate a finite amount of letter combinations, there are situations where not every link will be assigned a letter combination. See issue #65 for discussion.

The current solution is to filter out links without letter assignments before displaying popovers, however because we're overwriting the letters in LivePreviewLinkProcessor there are still certain situations where the links do not have an assigned letter combination and the popover will display undefined.

The solution is to rerun the filter in LivePreviewLinkProccessor after overwriting.